# Vectorise BitArray for ARM64