@Gnbrkm41 2020-03-19
Is there a way to check which tests specifically have failed & their console outputs? `Libraries Test Run release mono Linux arm64 Debug` is the only ARM test leg that had run & is reporting failures, but it does not actually specify what has failed (AzDO tests page just say "The work item failed") and the logs don't help either.