@tannergooding 2020-03-19
@Gnbrkm41, its not super intuitive, but once you are at the Azure logs (click `Details` on the failed job then `View more details on Azure Pipelines`), you can click on the build number at the top of the screen (in this case `20200319.8`) and then change from `Summary` to `Tests` which will give you:
![image](https://user-images.githubusercontent.com/10487869/77083712-29d75480-69bb-11ea-9e6c-2f2a9951b2ab.png)

You then have to click on the test, goto `Attachments`, and download the log.
-- Its worth noting it isn't always this complex. I think this is just because normal unit test results couldn't be reported.

In this case, the log indicates:
```
exit code 139 means SIGSEGV Illegal memory access. Deref invalid pointer, overrunning buffer, stack overflow etc. Core dumped.
```