@tannergooding 2020-03-19
Given that this is a core dump, I'll try to build and run locally to see if I can repro.