@echesakovMSFT 2020-03-25
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Snapdragon 835, 1 CPU, 8 logical and 8 physical cores
.NET Core SDK=5.0.100-preview.3.20174.2
  [Host]     : .NET Core 5.0.0 (CoreCLR 5.0.20.16901, CoreFX 5.0.20.16901), Arm64 RyuJIT
  Job-EZFDDC : .NET Core 5.0.0 (CoreCLR 42.42.42.42424, CoreFX 42.42.42.42424), Arm64 RyuJIT
  Job-CCIETN : .NET Core 5.0.0 (CoreCLR 42.42.42.42424, CoreFX 42.42.42.42424), Arm64 RyuJIT

PowerPlanMode=00000000-0000-0000-0000-000000000000  Arguments=/p:DebugType=portable  IterationTime=250.0000 ms  
MaxIterationCount=20  MinIterationCount=15  WarmupCount=1  

```
|                  Method |                                                                                 Toolchain | Size |          Mean |       Error |      StdDev |        Median |           Min |           Max | Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------ |------------------------------------------------------------------------------------------ |----- |--------------:|------------:|------------:|--------------:|--------------:|--------------:|------:|--------:|-------:|------:|------:|----------:|
|                         |                                                                                           |      |               |             |             |               |               |               |       |         |        |       |       |           |
|   BitArrayBoolArrayCtor | \884d68bf40f\artifacts\tests\coreclr\Windows_NT.arm64.Release\Tests\Core_Root\CoreRun.exe |  512 |    247.936 ns |   6.9971 ns |   8.0578 ns |    245.744 ns |    235.151 ns |    261.766 ns |  0.09 |    0.00 | 0.1147 |     - |     - |     120 B |
|   BitArrayBoolArrayCtor |    \baseline\artifacts\tests\coreclr\Windows_NT.arm64.Release\Tests\Core_Root\CoreRun.exe |  512 |  2,810.359 ns |  29.2818 ns |  27.3902 ns |  2,805.749 ns |  2,767.053 ns |  2,881.121 ns |  1.00 |    0.00 | 0.0665 |     - |     - |     120 B |
|                         |                                                                                           |      |               |             |             |               |               |               |       |         |        |       |       |           |
|             BitArrayNot | \884d68bf40f\artifacts\tests\coreclr\Windows_NT.arm64.Release\Tests\Core_Root\CoreRun.exe |  512 |     86.890 ns |   0.5803 ns |   0.5428 ns |     87.127 ns |     85.888 ns |     87.801 ns |  0.39 |    0.00 |      - |     - |     - |         - |
|             BitArrayNot |    \baseline\artifacts\tests\coreclr\Windows_NT.arm64.Release\Tests\Core_Root\CoreRun.exe |  512 |    224.989 ns |   1.3396 ns |   1.2531 ns |    225.491 ns |    222.566 ns |    226.437 ns |  1.00 |    0.00 |      - |     - |     - |         - |
|                         |                                                                                           |      |               |             |             |               |               |               |       |         |        |       |       |           |
|             BitArrayAnd | \884d68bf40f\artifacts\tests\coreclr\Windows_NT.arm64.Release\Tests\Core_Root\CoreRun.exe |  512 |    104.489 ns |   0.7997 ns |   0.7480 ns |    104.777 ns |    103.136 ns |    105.787 ns |  0.32 |    0.00 |      - |     - |     - |         - |
|             BitArrayAnd |    \baseline\artifacts\tests\coreclr\Windows_NT.arm64.Release\Tests\Core_Root\CoreRun.exe |  512 |    327.520 ns |   2.0465 ns |   1.9143 ns |    327.426 ns |    324.433 ns |    330.204 ns |  1.00 |    0.00 |      - |     - |     - |         - |
|                         |                                                                                           |      |               |             |             |               |               |               |       |         |        |       |       |           |
|              BitArrayOr | \884d68bf40f\artifacts\tests\coreclr\Windows_NT.arm64.Release\Tests\Core_Root\CoreRun.exe |  512 |    104.845 ns |   0.6562 ns |   0.6139 ns |    104.846 ns |    103.688 ns |    105.664 ns |  0.32 |    0.00 |      - |     - |     - |         - |
|              BitArrayOr |    \baseline\artifacts\tests\coreclr\Windows_NT.arm64.Release\Tests\Core_Root\CoreRun.exe |  512 |    328.054 ns |   1.6866 ns |   1.5777 ns |    327.635 ns |    325.572 ns |    331.016 ns |  1.00 |    0.00 |      - |     - |     - |         - |
|                         |                                                                                           |      |               |             |             |               |               |               |       |         |        |       |       |           |
|             BitArrayXor | \884d68bf40f\artifacts\tests\coreclr\Windows_NT.arm64.Release\Tests\Core_Root\CoreRun.exe |  512 |    104.698 ns |   0.6160 ns |   0.5762 ns |    104.708 ns |    103.509 ns |    105.599 ns |  0.32 |    0.00 |      - |     - |     - |         - |
|             BitArrayXor |    \baseline\artifacts\tests\coreclr\Windows_NT.arm64.Release\Tests\Core_Root\CoreRun.exe |  512 |    327.789 ns |   1.3957 ns |   1.3055 ns |    327.620 ns |    325.926 ns |    330.182 ns |  1.00 |    0.00 |      - |     - |     - |         - |
|                         |                                                                                           |      |               |             |             |               |               |               |       |         |        |       |       |           |
| BitArrayCopyToBoolArray | \884d68bf40f\artifacts\tests\coreclr\Windows_NT.arm64.Release\Tests\Core_Root\CoreRun.exe |  512 |  2,330.797 ns |   7.1107 ns |   6.3035 ns |  2,331.367 ns |  2,319.311 ns |  2,341.284 ns |  0.12 |    0.00 |      - |     - |     - |         - |
| BitArrayCopyToBoolArray |    \baseline\artifacts\tests\coreclr\Windows_NT.arm64.Release\Tests\Core_Root\CoreRun.exe |  512 | 19,189.121 ns | 126.9124 ns | 118.7140 ns | 19,168.260 ns | 19,012.465 ns | 19,360.812 ns |  1.00 |    0.00 |      - |     - |     - |         - |


[System.Collections.Tests.Perf_BitArray-report.zip](https://github.com/dotnet/runtime/files/4378609/System.Collections.Tests.Perf_BitArray-report.zip)